refactor: move check invariant in lib
